requirements.txt
Django >=2.1.3,<=2.2.0
requests == 2.18.4
git+https://github.com/openimis/openhim-mediator-utils-py
djangorestframework == 3.10.3
python-dotenv==0.17.0